feat: use WP_HTML_Tag_Processor for AttributesOutputFilter
